refactor: move Agent from theoriq/ to theoriq/api/v1alpha2
