+++ theoriq/__init__.py
@@ -1,4 +1,3 @@
-from .agent import AgentDeploymentConfiguration, Agent
-
+from .api.v1alpha2.agent import AgentDeploymentConfiguration, Agent
 from .api.v1alpha2.execute import ExecuteContext, ExecuteResponse
 from .api.common import ExecuteRuntimeError
