+++ theoriq/extra/globals.py
@@ -1,6 +1,6 @@
 from contextvars import ContextVar
 
-from theoriq.agent import Agent
+from theoriq.api.v1alpha2.agent import Agent
 
 # Global variable used to access the current agent context
 agent_var: ContextVar[Agent] = ContextVar("agent")
